requirements.txt
django
django-cors-headers
django-oauth-toolkit
django-rest-framework-mongoengine
djangorestframework
djangorestframework-simplejwt
mongoengine
channels
channels-redis
daphne